+++ tests/retime/test_zero_motions.py
@@ -8,11 +8,12 @@ import toppra
 
 toppra.setup_logging(level="INFO")
 
+# TODO:It's better to simple return failure whenever zero motion is
+# detected. This will belongs to the case of ERR_NUMERICAL_DIFFICULTY
 
-@pytest.mark.skip(reason="scaling is deprecated")
-@pytest.mark.parametrize("scaling", [1e-3, 1e-4])
-@pytest.mark.parametrize("Ngrid", [101, 501, 1001])
-def test_scalar_zero_motion(scaling, Ngrid):
+@pytest.mark.skip(reason="Do not handle zero motion case.")
+@pytest.mark.parametrize("Ngrid", [101, 1001])
+def test_scalar_zero_motion(Ngrid):
     """The simple zero motion trajectory
 
     Note: Paths with very small displacement, like the one given in
@@ -37,14 +38,14 @@ def test_scalar_zero_motion(scaling, Ngrid):
 
     instance = toppra.algorithm.TOPPRA(
         [pc_vel, pc_acc], path, solver_wrapper='hotqpoases',
-        gridpoints=np.linspace(0, 1.0, Ngrid), scaling=scaling)
-    jnt_traj, aux_traj, data = instance.compute_trajectory(0, 0, return_data=True)
+        gridpoints=np.linspace(0, 1.0, Ngrid))
+    jnt_traj = instance.compute_trajectory(0, 0, return_data=True)
     # Simply assert success
     assert jnt_traj is not None
     assert jnt_traj.duration < 9e-4  # less than 1ms
 
 
-@pytest.mark.skip(reason="scaling is deprecated")
+@pytest.mark.skip(reason="Do not handle zero motion case.")
 @pytest.mark.parametrize("Ngrid", [101, 501, 1001])
 def test_scalar_auto_scaling(Ngrid):
     """Automatic scaling should lead to better results at slower
@@ -65,8 +66,8 @@ def test_scalar_auto_scaling(Ngrid):
 
     instance = toppra.algorithm.TOPPRA(
         [pc_vel, pc_acc], path, solver_wrapper='hotqpoases',
-        gridpoints=np.linspace(0, 1.0, Ngrid), scaling=-1)
-    jnt_traj, aux_traj, data = instance.compute_trajectory(0, 0, return_data=True)
+        gridpoints=np.linspace(0, 1.0, Ngrid))
+    jnt_traj = instance.compute_trajectory(0, 0, return_data=True)
 
     # Simply assert success
     assert jnt_traj is not None